Make Facebook Gallery Picture Elements Accessible
Currently, the ft-more-photos container has photos being shown as `<a>` elements with a background image and no accompanying text (not as part of the link, title, or aria-label). Having links without any descriptors is an error when checking for whether something meets the established accessibility standards.

I've made it so the existing code for retrieving the picture description is made available (previously commented out) and then made it so that's then used as title & aria-label for the `<a>` elements showing the image(s). I also adjusted the fallback description so it always has some descriptor text (using a blank label if a description isn't available really doesn't resolve the accessibility issue.)